Fix docker PID 1 zombie reaping issue.
Unix processes normally follow a tree structure, where the init process
is the root of the tree. When a child process terminates, it keeps
occupying a slot in the process table, along with its exit status as a
so-called zombie process, until it has been waited for by its parent
process.

If the parent process terminates without ever having waited for the
child process, the child process is reparented under the root of the
process tree, i.e. the process with PID 1, which then becomes
responsible for waiting for ("reaping") zombie processes.

On a normal Unix system, that process is the init process. On a docker
system, that process is whatever is specified by CMD or ENTRYPOINT or
given on the command line. If that process does not reap zombie
processes, they will eventually flood the system.

Absent an init process, the next best choice is bash, which also takes
care of that. As long as bash is the PID 1 process, zombie processes
will be properly reaped.

This means that a docker image should ideally start by invoking a shell
or by executing a shell script. That shell script should not "exec"
another process at the end, or the top level process will be replaced by
the exec'ed process.

We used to do exactly that, exec'ing jenkins from the PID 1 process or
calling it directly, which led to zombie processes proliferating on the
host.