Manage Jenkins jobs as separate docker processes.
This is a major reworking of the CI logic. Jenkins now runs
outside docker and only executes build and test steps inside
separate docker images. This allows us to cleanly separate
instances so that running multiple concurrent jobs does not
lead to conflicts e.g. when accessing the home directory.