Introduce a setting to cap the memory usage of tests.
Some tests may occasionally use large amounts of memory due to
randomization. For our more memory-constrained test servers, this
setting allows us to limit the memory they use.

Note that running into that limit will generally result in a test
failure. However, it should be clear from the log that the failure
is due to an out of memory condition.